+++ webapps/smsc/WEB-INF/src/ru/novosoft/smsc/admin/smsstat/HourCountersSet.java
@@ -17,16 +17,14 @@ public class HourCountersSet extends CountersSet
   {
     this.hour = hour;
   }
-
-  public HourCountersSet(int accepted, int rejected, int delivered, int failed,
-                         int rescheduled, int temporal, int i, int o, int hour)
+  public HourCountersSet(long accepted, long  rejected, long delivered, long failed,
+                         long rescheduled, long temporal, long i, long o, int hour)
   {
     super(accepted, rejected, delivered, failed, rescheduled, temporal, i, o);
     this.hour = hour;
   }
 
-  public int getHour()
-  {
+  public int getHour() {
     return hour;
   }
 }
